requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
youtube-transcript-api>=0.6.2
google-api-python-client>=2.125.0
python-dotenv>=1.0.1
faiss-cpu>=1.7.4
